Refactor: Row ID (#120)
* Refactor: block id is removed and the block id info can be calculated from segment_offset of RowID


* Refactor: block capacity/row shouldbe u16 type, segment id should be u32 type


---------

+++ src/storage/wal/wal_entry.cpp
@@ -49,11 +49,11 @@ SharedPtr<WalCmd> WalCmd::ReadAdv(char *&ptr, i32 max_bytes) {
             String db_name = ReadBufAdv<String>(ptr);
             String table_name = ReadBufAdv<String>(ptr);
             String segment_dir = ReadBufAdv<String>(ptr);
-            i32 segment_id = ReadBufAdv<i32>(ptr);
-            i32 block_entries_size = ReadBufAdv<i32>(ptr);
-            Vector<i32> row_counts;
-            for (i32 i = 0; i < block_entries_size; ++i) {
-                auto row_count = ReadBufAdv<i32>(ptr);
+            u32 segment_id = ReadBufAdv<u32>(ptr);
+            u16 block_entries_size = ReadBufAdv<u16>(ptr);
+            Vector<u16> row_counts;
+            for (u16 i = 0; i < block_entries_size; ++i) {
+                auto row_count = ReadBufAdv<u16>(ptr);
                 row_counts.push_back(row_count);
             }
             cmd = MakeShared<WalCmdImport>(db_name, table_name, segment_dir, segment_id, block_entries_size, row_counts);
@@ -182,7 +182,7 @@ i32 WalCmdDropIndex::GetSizeInBytes() const {
 
 i32 WalCmdImport::GetSizeInBytes() const {
     return sizeof(WalCommandType) + sizeof(i32) + this->db_name.size() + sizeof(i32) + this->table_name.size() + sizeof(i32) +
-           this->segment_dir.size() + sizeof(i32) + sizeof(i32) + this->row_counts_.size() * sizeof(i32);
+           this->segment_dir.size() + sizeof(u32) + sizeof(u16) + this->row_counts_.size() * sizeof(u16);
 }
 
 i32 WalCmdAppend::GetSizeInBytes() const {
